Moved AnalysisCheck to another folder.
